Move mocks deploy to separate script (#26)
* fix: Move mocks deployment to seperate script

* fix: Run formatter

+++ src/Interfaces/IBondCalculator.sol
@@ -3,6 +3,5 @@
 pragma solidity 0.7.5;
 
 interface IBondCalculator {
-  function valuation( address pair_, uint amount_ ) external view returns ( uint _value );
+    function valuation(address pair_, uint256 amount_) external view returns (uint256 _value);
 }
-
